+++ skeletons/INTEGER.h
@@ -18,7 +18,7 @@ extern asn_TYPE_descriptor_t asn_DEF_INTEGER;
 
 /* Map with <tag> to integer value association */
 typedef struct asn_INTEGER_enum_map_s {
-	int64_t		 nat_value;	/* associated native integer value */
+	long long	 nat_value;	/* associated native integer value */
 	size_t		 enum_len;	/* strlen("tag") */
 	const char	*enum_name;	/* "tag" */
 } asn_INTEGER_enum_map_t;
@@ -56,12 +56,12 @@ per_type_encoder_f INTEGER_encode_aper;
  */
 int asn_INTEGER2int64(const INTEGER_t *i, int64_t *l);
 int asn_INTEGER2uint64(const INTEGER_t *i, uint64_t *l);
-int asn_INTEGER2long(const INTEGER_t *i, long *l);
-int asn_INTEGER2ulong(const INTEGER_t *i, unsigned long *l);
+int asn_INTEGER2long(const INTEGER_t *i, long long *l);
+int asn_INTEGER2ulong(const INTEGER_t *i, unsigned long long *l);
 int asn_int642INTEGER(INTEGER_t *i, int64_t l);
 int asn_uint642INTEGER(INTEGER_t *i, uint64_t l);
-int asn_long2INTEGER(INTEGER_t *i, long l);
-int asn_ulong2INTEGER(INTEGER_t *i, unsigned long l);
+int asn_long2INTEGER(INTEGER_t *i, long long l);
+int asn_ulong2INTEGER(INTEGER_t *i, unsigned long long l);
 
 /* A a reified version of strtol(3) with nicer error reporting. */
 enum asn_strtol_result_e {
@@ -71,15 +71,15 @@ enum asn_strtol_result_e {
     ASN_STRTOL_OK          =  0,  /* Conversion succeded, number ends at (*end) */
     ASN_STRTOL_EXTRA_DATA  =  1   /* Conversion succeded, but the string has extra stuff */
 };
-enum asn_strtol_result_e asn_strtol_lim(const char *str, const char **end, long *l);
+enum asn_strtol_result_e asn_strtol_lim(const char *str, const char **end, long long *l);
 
 /* The asn_strtol is going to be DEPRECATED soon */
-enum asn_strtol_result_e asn_strtol(const char *str, const char *end, long *l);
+enum asn_strtol_result_e asn_strtol(const char *str, const char *end, long long *l);
 
 /*
  * Convert the integer value into the corresponding enumeration map entry.
  */
-const asn_INTEGER_enum_map_t *INTEGER_map_value2enum(asn_INTEGER_specifics_t *specs, long value);
+const asn_INTEGER_enum_map_t *INTEGER_map_value2enum(asn_INTEGER_specifics_t *specs, long long value);
 
 #ifdef __cplusplus
 }
